Casi todo listo!

+++ backoffice-Back-End/src/main/java/edu/bootcamp/backoffice/model/orderDetail/productDetail/dto/ProductDetailRequest.java
@@ -16,5 +16,5 @@ public class ProductDetailRequest {
   private Integer quantity;
 
   @JsonProperty("warranty")
-  private Float warranty;
+  private Double warranty;
 }
